+++ pyoupyou/urls.py
@@ -63,6 +63,7 @@ urlpatterns = [
     url(r"^create_offer/$", views.create_offer_ajax, name="create_offer"),
     url(r"^feed/pyoupyou_full.ics$", feeds.InterviewFeed(), name="calendar_full"),
     url(r"^export/all_interviews.tsv$", views.export_interviews_tsv, name="export_interviews_tsv"),
+    url(r"^export/all_processes.tsv$", views.export_processes_tsv, name="export_processess_tsv"),
     url(r"^select2/", include("django_select2.urls")),
     url(r"^search/", views.search, name="search"),
 ]