+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.5.0 ()
+- Speed up interviews export
+- Add processes export
+
 # 1.4.0 (2020-05-14)
 - Add a page listing processes for a given source